+++ testsuite/modern-testsuite/src/test/java/categories/npm.java
@@ -10,11 +10,11 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.testsuite.npm;
+package categories;
 
 /**
  * Junit4 category of NPM related ITs.
  */
-public interface NpmCategory
+public interface npm
 {
 }
