+++ testsuite/modern-testsuite/src/test/java/org/sonatype/nexus/testsuite/npm/NpmITSupport.java
@@ -22,6 +22,7 @@ import org.sonatype.nexus.client.core.subsystem.repository.Repositories;
 import org.sonatype.nexus.testsuite.support.NexusRunningParametrizedITSupport;
 import org.sonatype.nexus.testsuite.support.NexusStartAndStopStrategy;
 
+import categories.npm;
 import com.bolyuba.nexus.plugin.npm.client.NpmGroupRepository;
 import com.bolyuba.nexus.plugin.npm.client.NpmHostedRepository;
 import com.bolyuba.nexus.plugin.npm.client.NpmProxyRepository;
@@ -43,7 +44,7 @@ import static org.sonatype.sisu.goodies.common.Varargs.$;
 /**
  * Support for NPM integration tests.
  */
-@Category(NpmCategory.class)
+@Category(npm.class)
 @NexusStartAndStopStrategy(EACH_TEST)
 public abstract class NpmITSupport
     extends NexusRunningParametrizedITSupport
@@ -65,6 +66,7 @@ public abstract class NpmITSupport
   protected NexusBundleConfiguration configureNexus(final NexusBundleConfiguration configuration) {
     return super.configureNexus(configuration)
         .setLogLevel("com.bolyuba.nexus.plugin.npm", "DEBUG")
+        .setLogLevel("remote.storage.outbound", "DEBUG") // see outbound HTTP requests
         .addPlugins(
             artifactResolver().resolvePluginFromDependencyManagement(
                 "org.sonatype.nexus.plugins", "nexus-npm-repository-plugin"
